requirements.txt
pytest
requests
allure-pytest
colorlog
pytest-xdist
pytest-rerunfailures
pyyaml